+++ packages/timeline/src/components/Timeline/Timeline.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useRef, useState } from 'react'
 import c from './Timeline.module.css'
 import '@hedron-gl/ui-core/base.css'
 import '@hedron-gl/ui-core/fonts.css'
-import type { Timeline as TimelineData } from './types'
+import type { Timeline as TimelineData } from '@types'
 
 export interface TimelineProps {
   /** Timeline data */
